+++ include/kImageAnnotator/KImageAnnotator.h
@@ -22,7 +22,6 @@
 
 #include <QWidget>
 
-#include "NumberUpdateMode.h"
 #include <kImageAnnotator/KImageAnnotatorExport.h>
 
 namespace kImageAnnotator {
@@ -62,7 +61,7 @@ public Q_SLOTS:
 	void setSaveToolSelection(bool enabled);
 	void setSmoothFactor(int factor);
 	void setSwitchToSelectToolAfterDrawingItem(bool enabled);
-	void setNumberUpdateMode(NumberUpdateMode numberUpdateMode);
+	void setNumberToolSeedChangeUpdatesAllItems(bool enabled);
 	void setTabBarAutoHide(bool enabled);
 	void removeTab(int index);
 	void setStickers(const QStringList &stickerPaths, bool keepDefault);